[refactor] 코드 구조 개선
